Rename sample module package name
