+++ sample/src/main/java/jp/co/tokubai/Page.java
@@ -1,0 +1,21 @@
+package jp.co.tokubai;
+
+public class Page {
+
+    private final int text;
+
+    private final int subtitle;
+
+    public Page(int subtitle, int text) {
+        this.subtitle = subtitle;
+        this.text = text;
+    }
+
+    public int getText() {
+        return text;
+    }
+
+    public int getSubtitle() {
+        return subtitle;
+    }
+}
