+++ sample/src/main/java/jp/co/tokubai/extension/ExtensionPinFragment.java
@@ -1,0 +1,33 @@
+package jp.co.tokubai.extension;
+
+import android.graphics.PointF;
+import android.os.Bundle;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import jp.co.tokubai.ImageSource;
+import jp.co.tokubai.subscaleview.test.R.id;
+import jp.co.tokubai.subscaleview.test.R.layout;
+import jp.co.tokubai.extension.views.PinView;
+
+public class ExtensionPinFragment extends Fragment {
+
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        View rootView = inflater.inflate(layout.extension_pin_fragment, container, false);
+        final ExtensionActivity activity = (ExtensionActivity)getActivity();
+        if (activity != null) {
+            rootView.findViewById(id.next).setOnClickListener(new View.OnClickListener() {
+                @Override public void onClick(View v) { activity.next(); }
+            });
+        }
+        PinView imageView = rootView.findViewById(id.imageView);
+        imageView.setImage(ImageSource.asset("sanmartino.jpg"));
+        imageView.setPin(new PointF(1602f, 405f));
+        return rootView;
+    }
+
+}
