+++ sample/src/main/res/layout/animation_activity.xml
@@ -19,6 +19,7 @@
 
     </RelativeLayout>
 
-    <com.davemorrissey.labs.subscaleview.test.extension.views.PinView style="@style/mainPanel" android:id="@+id/imageView"/>
+    <jp.co.tokubai.extension.views.PinView
+        style="@style/mainPanel" android:id="@+id/imageView"/>
 
 </RelativeLayout>
