+++ sample/src/main/res/layout/extension_circle_fragment.xml
@@ -14,6 +14,7 @@
 
     </RelativeLayout>
 
-    <com.davemorrissey.labs.subscaleview.test.extension.views.CircleView style="@style/mainPanel" android:id="@+id/imageView"/>
+    <jp.co.tokubai.extension.views.CircleView
+        style="@style/mainPanel" android:id="@+id/imageView"/>
 
 </RelativeLayout>
