+++ sample/src/main/res/layout/extension_freehand_fragment.xml
@@ -20,6 +20,7 @@
 
     </RelativeLayout>
 
-    <com.davemorrissey.labs.subscaleview.test.extension.views.FreehandView style="@style/mainPanel" android:id="@+id/imageView"/>
+    <jp.co.tokubai.extension.views.FreehandView
+        style="@style/mainPanel" android:id="@+id/imageView"/>
 
 </RelativeLayout>
