+++ sample/src/main/res/layout/view_pager.xml
@@ -15,6 +15,7 @@
 
     <androidx.viewpager.widget.ViewPager style="@style/mainPanel" android:id="@+id/horizontal_pager" android:visibility="gone" />
 
-    <com.davemorrissey.labs.subscaleview.test.viewpager.VerticalViewPager style="@style/mainPanel" android:id="@+id/vertical_pager" android:visibility="gone"/>
+    <jp.co.tokubai.viewpager.VerticalViewPager
+        style="@style/mainPanel" android:id="@+id/vertical_pager" android:visibility="gone"/>
 
 </RelativeLayout>
